refactor: extract ClaudeEventHandler, make EventHandler an orchestrator
